refactor: split RequestPanels and BannerGroup into individual components
RequestPanels.ts (1025 → 272 lines): Split 4 distinct panel types into
individual components (ToolEditRequestPanel, BashRequestPanel,
RetryRequestPanel, ProposalRequestPanel) with a BaseFeedbackPanel base
class for shared feedback/reject/emit logic. The container now handles
only permission grouping, section headers, and keyboard shortcut delegation.

BannerGroup.ts (451 → 71 lines): Split 5 banner types into individual
components (ApiKeyBanner, AgentConfigBanner, DependencyBanner,
GettingStartedBanner, LoginBanner) with shared warningBannerStyles.
The container now only passes through state properties.

Each component is independently testable with its own rendering, state,
and event handling. No changes to external interfaces — both
<request-panels> and <banner-group> custom elements remain unchanged.
