+++ src/webview/frontend/components/AgentConfigBanner.ts
@@ -1,0 +1,103 @@
+/**
+ * Banner component for missing agent configuration.
+ *
+ * Displays a warning banner when agent configuration is missing,
+ * with actions to edit agents, set directory, or view docs.
+ *
+ * @fires agent-config-action - When edit/dir/docs button is clicked
+ */
+
+// Third-party imports
+import { LitElement, html, nothing, type TemplateResult } from 'lit';
+import { customElement, property } from 'lit/decorators.js';
+
+// Local imports - shared styles
+import { designTokens, codiconStyles, commonViewStyles } from '@shared/styles';
+
+// Local imports - shared schemas
+
+// Local imports - shared banner styles
+import { warningBannerStyles } from '../styles/warningBannerStyles';
+
+// Local imports - main view events
+import { MainViewEvents } from '../events';
+import type { AgentConfigBannerState } from '@shared/schemas';
+
+@customElement('agent-config-banner')
+export class AgentConfigBanner extends LitElement {
+  static override styles = [
+    designTokens,
+    codiconStyles,
+    commonViewStyles,
+    warningBannerStyles,
+  ];
+
+  @property({ attribute: false }) state: AgentConfigBannerState = {
+    visible: false,
+  };
+
+  private handleActionClick(event: MouseEvent): void {
+    const button = (event.target as HTMLElement).closest<HTMLElement>(
+      '[data-action]',
+    );
+    const action = button?.dataset.action as
+      | 'edit'
+      | 'dir'
+      | 'docs'
+      | undefined;
+    if (!action) return;
+    this.dispatchEvent(
+      MainViewEvents.agentConfigAction({
+        action,
+        customDirSet: this.state.customDirSet,
+      }),
+    );
+  }
+
+  override render(): TemplateResult | typeof nothing {
+    if (!this.state.visible) return nothing;
+
+    return html`
+      <div
+        id="agentConfigBanner"
+        class="warning-banner"
+        data-custom-dir-set=${this.state.customDirSet ? 'true' : 'false'}
+      >
+        <span>
+          ${this.state.agentName
+            ? `Agent file for "${this.state.agentName}" is missing.`
+            : 'Agent configuration is missing.'}
+        </span>
+        <div class="actions" @click=${this.handleActionClick}>
+          <vscode-toolbar-button
+            id="agentConfigEditButton"
+            icon="edit"
+            data-action="edit"
+          >
+            Edit Agents
+          </vscode-toolbar-button>
+          <vscode-toolbar-button
+            id="agentConfigDirButton"
+            icon="folder"
+            data-action="dir"
+          >
+            ${this.state.customDirSet ? 'Open Directory' : 'Set Directory'}
+          </vscode-toolbar-button>
+          <vscode-toolbar-button
+            id="agentConfigDocButton"
+            icon="book"
+            data-action="docs"
+          >
+            Docs
+          </vscode-toolbar-button>
+        </div>
+      </div>
+    `;
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'agent-config-banner': AgentConfigBanner;
+  }
+}
